Move comment section styles to new file
