+++ instagram-david/src/components/CommentSection/CommentSection.js
@@ -1,49 +1,18 @@
 import React from "react";
-import styled from "styled-components";
 import PropTypes from "prop-types";
-import Comment from "./Comment";
+import Comment from "./Comment/Comment";
 
-const CommentSectionContainer = styled.div`
-  padding: 15px;
-`;
-const LikesCommentsIcons = styled.div`
-  margin-bottom: 15px;
-`;
-const I = styled.div`
-  margin-right: 20px;
-`;
-const LikesCommentsStats = styled.div`
-  font-size: 1.5rem;
-  font-weight: 700;
-  margin-bottom: 10px;
-`;
-const LikesCommentsStatistic = styled.p`
-  margin-bottom: 5px;
-`;
-const CommentForm = styled.form`
-  width: 95%;
-  margin: 10px auto 0px;
-  display: flex;
-  justify-content: space-between;
-`;
-const CommentInput = styled.input`
-  width: 80%;
-  padding: 10px;
-  border: none;
-`;
-const SubmitComment = styled.button`
-  width: 100px;
-  color: white;
-  background-color: #3897f0;
-  border: none;
-  outline: none;
-  border-radius: 5px;
-  padding: 10px;
-  font-size: 1.5rem;
-  :active {
-    transform: scale(0.99);
-  }
-`;
+import {
+  CommentSectionContainer,
+  LikesCommentsIcons,
+  Icon,
+  LikesCommentsStats,
+  LikesCommentsStatistic,
+  CommentForm,
+  CommentInput,
+  SubmitComment
+} from "./CommentSectionStyles";
+
 class CommentSection extends React.Component {
   constructor(props) {
     super(props);
@@ -109,8 +78,8 @@ class CommentSection extends React.Component {
       <CommentSectionContainer>
         <div>
           <LikesCommentsIcons>
-            <I className="far fa-heart fa-2x" onClick={this.toggleLike} />
-            <I
+            <Icon className="far fa-heart fa-2x" onClick={this.toggleLike} />
+            <Icon
               className="far fa-comment fa-2x"
               onClick={this.commentInputFocus}
             />
